fix: seperate component of instance pages
